+++ internal/provider/agent_pool_token_resource_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/plancheck"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	"github.com/scalr/go-scalr"
 )
@@ -62,6 +63,42 @@ func TestAccScalrAgentPoolToken_update(t *testing.T) {
 					testAccCheckScalrAgentPoolTokenExists("scalr_agent_pool_token.test", pool, token),
 					resource.TestCheckResourceAttr("scalr_agent_pool_token.test", "description", "agent_pool_token-updated"),
 				),
+			},
+		},
+	})
+}
+
+func TestAccScalrAgentPoolToken_UpgradeFromSDK(t *testing.T) {
+	var pool scalr.AgentPool
+	if isAccTest() {
+		pool = createPool(t)
+		defer deletePool(t, pool)
+	}
+
+	resource.Test(t, resource.TestCase{
+		Steps: []resource.TestStep{
+			{
+				ExternalProviders: map[string]resource.ExternalProvider{
+					"scalr": {
+						Source:            "registry.scalr.io/scalr/scalr",
+						VersionConstraint: "<=2.4.0",
+					},
+				},
+				Config: testAccScalrAgentPoolTokenBasic(pool),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("scalr_agent_pool_token.test", "description", "agent_pool_token-test"),
+					resource.TestCheckResourceAttr("scalr_agent_pool_token.test", "agent_pool_id", pool.ID),
+					resource.TestCheckResourceAttrSet("scalr_agent_pool_token.test", "token"),
+				),
+			},
+			{
+				ProtoV5ProviderFactories: protoV5ProviderFactories(t),
+				Config:                   testAccScalrAgentPoolTokenBasic(pool),
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{
+						plancheck.ExpectEmptyPlan(),
+					},
+				},
 			},
 		},
 	})